chore: Made Host Listing Sidebar reusable for all category
