+++ apps/web/app/(listing-activity)/layout.tsx
@@ -8,15 +8,16 @@ import React from "react"
 import { LOGO_SINGLE_IMAGE } from "@/common/constants/index"
 import { APP_NAME } from "@repo/constants"
 import Header from "@/common/components/HostLayout/ListingActivityHeader"
-import Sidebar from "@/common/components/HostLayout/ListingActivitySidebar"
+import ListingSidebar from "@/common/components/HostLayout/ListingSidebar"
 import AuthStateProvider from "@/common/components/AuthStateProvider"
+import { E_Listing_Category } from "@repo/contract"
 
 export const metadata: Metadata = {
   title: APP_NAME,
   description: `Generated by ${APP_NAME}`,
 }
 
-export default async function HostListingLayout({
+export default async function HostListingActivityLayout({
   children,
 }: {
   readonly children: React.ReactNode
@@ -30,7 +31,7 @@ export default async function HostListingLayout({
           <AuthStateProvider>
             <GlobalModalWrapper>
               <Header />
-              <Sidebar>{children}</Sidebar>
+              <ListingSidebar category={E_Listing_Category.Activity}>{children}</ListingSidebar>
             </GlobalModalWrapper>
           </AuthStateProvider>
         </QueryClientWrapper>
