+++ apps/web/app/(listing-property)/hosting/listings/properties/[listingId]/property-type/page.tsx
