+++ apps/web/app/(listing-property)/layout.tsx
@@ -7,16 +7,17 @@ import { Toaster } from "react-hot-toast"
 import React from "react"
 import { LOGO_SINGLE_IMAGE } from "@/common/constants/index"
 import { APP_NAME } from "@repo/constants"
-import Header from "@/common/components/HostLayout/NewListingHeader"
-import Sidebar from "@/common/components/HostLayout/NewListingSidebar"
+import Header from "@/common/components/HostLayout/ListingHeader"
+import ListingSidebar from "@/common/components/HostLayout/ListingSidebar"
 import AuthStateProvider from "@/common/components/AuthStateProvider"
+import { E_Listing_Category } from "@repo/contract"
 
 export const metadata: Metadata = {
   title: APP_NAME,
   description: `Generated by ${APP_NAME}`,
 }
 
-export default async function HostNewListingLayout({
+export default async function HostListingPropertyLayout({
   children,
 }: {
   readonly children: React.ReactNode
@@ -30,7 +31,7 @@ export default async function HostNewListingLayout({
           <AuthStateProvider>
             <GlobalModalWrapper>
               <Header />
-              <Sidebar>{children}</Sidebar>
+              <ListingSidebar category={E_Listing_Category.Property}>{children}</ListingSidebar>
             </GlobalModalWrapper>
           </AuthStateProvider>
         </QueryClientWrapper>
