+++ apps/web/app/(listing-rental)/hosting/listings/rentals/[listingId]/details/page.tsx
