+++ apps/web/common/components/AdminLayout/NewListingRentalSidebar.tsx
@@ -15,6 +15,7 @@ import NewListingEditIndicator from "./NewListingEditIndicator"
 import { BOTTOM_LINKS } from "@/common/constants"
 import { useParams } from "next/navigation"
 import useGetFinishedSections from "@/module/Hosting/Listings/hooks/useGetFinishedSections"
+import { E_Listing_Category } from "@repo/contract"
 
 interface HostSidebarProps {
   children: React.ReactNode
@@ -23,7 +24,7 @@ interface HostSidebarProps {
 const Sidebar = ({ children }: HostSidebarProps) => {
   const params = useParams<{ listingId: string }>()
   const listingId = String(params.listingId)
-  const { data } = useGetFinishedSections({ listingId, type: "rental" })
+  const { data } = useGetFinishedSections({ listingId, category: E_Listing_Category.Rental })
   const finishedSections = data?.item?.finishedSections || []
   const RENTAL_SETUP_BASE_PATH = "/hosting/listings/rentals/setup"
   const SECTION_LINKS = [
