+++ apps/web/common/components/AdminLayout/NewListingSidebar.tsx
@@ -17,6 +17,7 @@ import NewListingEditIndicator from "./NewListingEditIndicator"
 import useGetFinishedSections from "@/module/Hosting/Listings/hooks/useGetFinishedSections"
 import { useParams } from "next/navigation"
 import { BOTTOM_LINKS } from "@/common/constants"
+import { E_Listing_Category } from "@repo/contract"
 
 interface HostSidebarProps {
   children: React.ReactNode
@@ -25,7 +26,7 @@ interface HostSidebarProps {
 const Sidebar = ({ children }: HostSidebarProps) => {
   const params = useParams<{ listingId: string }>()
   const listingId = String(params.listingId)
-  const { data } = useGetFinishedSections({ listingId, type: "property" })
+  const { data } = useGetFinishedSections({ listingId, category: E_Listing_Category.Property })
   const finishedSections = data?.item?.finishedSections || []
   const PROPERTY_SETUP_BASE_PATH = "/hosting/listings/properties/setup"
   const SECTION_LINKS = [
